Removed unused functions and fixed SonarQube-identified issues. Extra values retrieved from single valued forecasts will now be added to the input rather than dropped.

+++ wres-io/src/wres/io/retrieval/BackToBackMetricInputIterator.java
@@ -246,7 +246,7 @@ final class BackToBackMetricInputIterator extends MetricInputIterator
             }
 
             double windowWidth = this.getProjectDetails().getWindowWidth();
-            double windowSpan = last - start;
+            double windowSpan = (double)(last - start);
 
             count = ((Double)Math.ceil( windowSpan / windowWidth)).intValue();
         }
